chatbot-python-service/requirements.txt
flask
faiss-cpu
sentence-transformers
nltk
mysql-connector-python
openpyxl
python-docx
pdfplumber
transformers
scikit-learn
tqdm
